# swagger.yaml
swagger: "2.0"
host: "marketo.swagger.io"
basePath: "/bulk/v1"
info:
  version: "1.0.0"
  title: "Marketo Bulk Export"
...
schemes:
- "https"
- "http"
paths:
...
  /activities/export/create.json:
    post:
      tags:
      - "Activities"
      summary: "Create Export Activity Job (Step 1)"
      description: "Create export job for search criteria defined via \"filter\" parameter. Request returns the \"exportId\" which is passed as a parameter in subsequent calls to Bulk Export Activities endpoints. Use Enqueue Export Activity Job endpoint to queue the export job for processing. Use Get Export Activity Job Status endpoint to retrieve status of export job. Required Permissions: Read-Only Activity"
      operationId: "exportActivityRequest"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Create new exportActivityRequest"
        required: true
        schema:
          $ref: "#/definitions/exportActivityRequest"
      responses:
        "405":
          description: "Invalid input"
      security:
      - marketo_auth:
        - "read:job"
...
securityDefinitions:
  marketo_auth:
    type: "oauth2"
    authorizationUrl: "https://marketo"
    flow: "implicit"
    scopes:
      write:job: "create job"
      read:job: "read job"
definitions:
  exportActivityRequest:
    type: "object"
    required:
    - "fields"
    - "filter"
    properties:
      columnHeaderNames:
        type: "object"
        required: 
        - "name"
        - "value"
        properties:
          name:
            type: "string"
          value:
            type: "string"
      fields:
        type: "array"
        items:
          type: "string"
      filter:
        type: "object"
        required:
        - "createdAt"
        properties:
          activityTypeIds:
            type: "array"
            items:
              type: "integer"
              format: "int64"
          primaryAttributeValueIds:
            type: "array"
            items:
              type: "integer"
              format: "int64"
          primaryAttributeValues:
            type: "array"
            items:
              type: "integer"
              format: "int64"
          createdAt:
            type: "object"
            properties:
              endAt: 
                type: "string"
              startAt:
                type: "string"
      format:
        type: "string"
      status:
        type: "string"
        description: "Export Job Status"
        enum:
        - "Created"
        - "Queued"
        - "Processing"
        - "Canceled"
        - "Completed"
        - "Failed"
    xml:
      name: "Export"
...